+++ src/app/terra-components.module.ts
@@ -87,6 +87,7 @@ import { TerraBaseService } from './service/terra-base.service';
 import { TerraUrlParamsDecorator } from './service/data/terra-url-params-decorator.service';
 import { TerraAlertComponent } from './components/alert/terra-alert.component';
 import { TerraTimePickerComponent } from './components/forms/input/time-picker/terra-time-picker.component';
+import { CategoryTreeConfig } from './components/category-picker/config/category-tree.config';
 import { TerraJsonToFormFieldService } from './components/forms/dynamic-form/service/terra-json-to-form-field.service';
 import { TerraDynamicFormComponent } from './components/forms/dynamic-form/terra-dynamic-form.component';
 import { TerraDynamicSwitchComponent } from './components/forms/dynamic-form/dynamic-switch/terra-dynamic-switch.component';
@@ -144,8 +145,11 @@ import { TerraMultiSplitViewBreadcrumbsService } from './components/split-view/m
 import { RouterModule } from '@angular/router';
 import { TerraTimePickerComponentExample } from './components/forms/input/time-picker/example/terra-time-picker.component.example';
 import { TerraTwoColumnsContainerComponent } from './components/layouts/column-container/terra-two-columns-container.component';
-import { CategoryTreeConfig } from './components/category-picker/config/category-tree.config';
 import { TerraTwoColumnsContainerDirective } from './components/layouts/column-container/terra-two-columns-container.directive';
+import { TerraBreadcrumbsComponent } from './components/breadcrumbs/terra-breadcrumbs.component';
+import { TerraBreadcrumbsService } from './components/breadcrumbs/service/terra-breadcrumbs.service';
+import { Type } from '@angular/core/src/type';
+import { DeviceDetectorModule } from 'ngx-device-detector';
 import { TerraTagSelectComponent } from './components/forms/tag-select/terra-tag-select.component';
 import { TerraTagSelectComponentExample } from './components/forms/tag-select/example/terra-tag-select.component.example';
 
@@ -218,6 +222,7 @@ function createCompiler(compilerFactory:CompilerFactory):Compiler
         TerraCategoryPickerComponent,
         TerraTwoColumnsContainerComponent,
         TerraTwoColumnsContainerDirective,
+        TerraBreadcrumbsComponent,
         TerraTagSelectComponent,
 
         // ### declarations of terra-components examples
@@ -316,6 +321,7 @@ function createCompiler(compilerFactory:CompilerFactory):Compiler
         TerraNodeComponent,
         TerraCategoryPickerComponent,
         TerraTwoColumnsContainerComponent,
+        TerraBreadcrumbsComponent,
         TerraTagSelectComponent
     ],
     exports:         [
@@ -378,6 +384,7 @@ function createCompiler(compilerFactory:CompilerFactory):Compiler
         TerraCategoryPickerComponent,
         TerraTwoColumnsContainerComponent,
         TerraTwoColumnsContainerDirective,
+        TerraBreadcrumbsComponent,
         TerraTagSelectComponent,
 
         // ### declarations of terra-components examples
@@ -443,7 +450,8 @@ function createCompiler(compilerFactory:CompilerFactory):Compiler
         AceEditorModule,
         TerraInteractModule,
         QuillModule,
-        RouterModule
+        RouterModule,
+        DeviceDetectorModule.forRoot()
     ],
     providers:       [
         TerraNavigatorSplitViewConfig,
@@ -492,29 +500,14 @@ export class TerraComponentsModule
                 TerraDynamicFormService,
                 TerraFormFieldControlService,
                 TerraJsonToFormFieldService,
-                TerraMultiSplitViewBreadcrumbsService
+                TerraMultiSplitViewBreadcrumbsService,
+                TerraBreadcrumbsService
             ]
         };
     }
 
-    public static forChild():ModuleWithProviders
+    public static forChild():Type<any>
     {
-        return {
-            ngModule:  TerraComponentsModule,
-            providers: [
-                TerraLoadingSpinnerService,
-                TerraDataTableContextMenuService,
-                TerraBaseService,
-                TerraNavigatorSplitViewConfig,
-                TerraUrlParamsDecorator,
-                TerraFrontendStorageService,
-                TerraAlertComponent,
-                TerraDynamicFormService,
-                CategoryTreeConfig,
-                TerraFormFieldControlService,
-                TerraJsonToFormFieldService,
-                TerraMultiSplitViewBreadcrumbsService
-            ]
-        };
+        return TerraComponentsModule;
     }
 }